+++ python/egglog/pretty.py
@@ -117,6 +117,10 @@ def pretty_callable_ref(
     return res[0] if isinstance(res, tuple) else res
 
 
+# TODO: Add a different pretty callable ref that doesnt fill in wholes but instead returns the function
+# so that things like Math.__add__ will be represented properly
+
+
 @dataclass
 class TraverseContext:
     """
@@ -170,6 +174,8 @@ class TraverseContext:
                 if until:
                     for f in until:
                         self(f)
+            case PartialCallDecl(c):
+                self(c)
             case _:
                 assert_never(decl)
 
@@ -231,6 +237,9 @@ class PrettyContext:
                 return name, name
             case CallDecl(_, _, _):
                 return self._call(decl, parens)
+            case PartialCallDecl(CallDecl(ref, typed_args, _)):
+                arg_strs = (_pretty_callable(ref), *(self(a.expr, parens=False, unwrap_lit=True) for a in typed_args))
+                return f"UnstableFn({', '.join(arg_strs)}", "fn"
             case PyObjectDecl(value):
                 return repr(value) if unwrap_lit else f"PyObject({value!r})", "PyObject"
             case ActionCommandDecl(action):
@@ -302,19 +311,28 @@ class PrettyContext:
             l, r = self(args[0]), self(args[1])
             return f"ne({l}).to({r})", "Unit"
         function_decl = self.decls.get_callable_decl(ref).to_function_decl()
+        signature = function_decl.signature
+
         # Determine how many of the last arguments are defaults, by iterating from the end and comparing the arg with the default
         n_defaults = 0
-        for arg, default in zip(
-            reversed(args), reversed(function_decl.arg_defaults), strict=not function_decl.var_arg_type
-        ):
-            if arg != default:
-                break
-            n_defaults += 1
+        # Dont try counting defaults for function application
+        if isinstance(signature, FunctionSignature):
+            for arg, default in zip(
+                reversed(args), reversed(signature.arg_defaults), strict=not signature.var_arg_type
+            ):
+                if arg != default:
+                    break
+                n_defaults += 1
         if n_defaults:
             args = args[:-n_defaults]
 
-        tp_name = function_decl.semantic_return_type.name
-        if function_decl.mutates:
+        # If this is a function application, the type is the first type arg of the function object
+        if signature == "fn-app":
+            tp_name = decl.args[0].tp.args[0].name
+        else:
+            assert isinstance(signature, FunctionSignature)
+            tp_name = signature.semantic_return_type.name
+        if isinstance(signature, FunctionSignature) and signature.mutates:
             first_arg = args[0]
             expr_str = self(first_arg)
             # copy an identifier expression iff it has multiple parents (b/c then we can't mutate it directly)
@@ -395,6 +413,28 @@ class PrettyContext:
             name = self._generate_name(tp_name)
             self.statements.append(f"{name} = {expr_str}")
         return name
+
+
+def _pretty_callable(ref: CallableRef) -> str:
+    """
+    Returns a function call as a string.
+    """
+    match ref:
+        case FunctionRef(name):
+            return name
+        case (
+            ClassMethodRef(class_name, method_name)
+            | MethodRef(class_name, method_name)
+            | PropertyRef(class_name, method_name)
+        ):
+            return f"{class_name}.{method_name}"
+        case ConstantRef(_):
+            msg = "Constants should not be callable"
+            raise NotImplementedError(msg)
+        case ClassVariableRef(_, _):
+            msg = "Class variables should not be callable"
+            raise NotADirectoryError(msg)
+    assert_never(ref)
 
 
 def _plot_line_length(expr: object):  # pragma: no cover